+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Rumor no longer clears the scope when starting
+- `Rumor.CallBinding` no longer calls DynamicInvoke and is much faster
 
 ### Fixed
 - Fix `Rumor.Choosing` throwing an error if Rumor has not been started